Rename `judgment_history` to `document_history`
This is a cosmetic filename change; in time all files will be renamed to
reflect the fact they're dealing with abstract `Document`s, and not
always `Judgment`s.
