+++ judgments/urls.py
@@ -5,6 +5,7 @@ from .views.button_handlers import (
     hold_judgment_button,
     prioritise_judgment_button,
 )
+from .views.document_history import DocumentHistoryView
 from .views.full_text import (
     html_view,
     html_view_redirect,
@@ -14,7 +15,6 @@ from .views.full_text import (
 )
 from .views.index import index
 from .views.judgment_edit import EditJudgmentView, edit_view_redirect
-from .views.judgment_history import DocumentHistoryView
 from .views.judgment_hold import (
     HoldJudgmentSuccessView,
     HoldJudgmentView,
